Admin Edit Product Gallery

+++ app/Http/Livewire/Admin/AdminProductComponent.php
@@ -19,7 +19,9 @@ class AdminProductComponent extends Component
 		if ($product->images) {
 			$images = explode(",", $product->images);
 			foreach ($images as $image) {
-				unlink('assets/images/products' . '/' . $image);
+				if ($image) {
+					unlink('assets/images/products' . '/' . $image);
+				}
 			}
 		}
 		$product->delete();
